// app/src/features/training/enums/SetType.ts
/** Role a prescribed set plays within an exercise. */
export enum SetType {
  Warmup = "warmup",
  Working = "working",
  TopSet = "top_set",
  BackOff = "back_off",
  DropSet = "drop_set",
  Amrap = "amrap",
  Failure = "failure",
  Myorep = "myorep",
  Cluster = "cluster",
  Deload = "deload",
}